Added runtime IM 2 setup detection
Even if IM 2 isn't active at the time of the snapshot, we can now detect
it being setup by the code we're tracing, so it can be tried later.

The I register and interrupt handler are now also checked for validity
before being used, in case the interrupt table hasn't yet been set up.