Added start of opcode tracking
In the output image, bright colours now indicate the start of each
instruction, which could be an extended set prefix.  Index prefixes
are included in addition to the base instruction, so LD IX,0 has
both DD and 21 flagged.

The default map output still includes operands, but the -z option
limits to just the start of instructions, for SkoolKit compatibility.