Fixed offset reading at the end of RAM
Intermediate address values (pc+x) may extend beyond the end of RAM,
accessing outside the array and potentially returning junk.  To fix
this the first 16 bytes of ROM are duplicated after the end of RAM.

This was needed when using the 0xff table in from for IM 2, where a
JR is placed at 0xffff with the offset being the F3 at the start of
the ROM.  This was previously working when we used pc-1 to pick out
the offset, but broken when changed to pc0+1.  Though the earlier
approach risked indexing back off the start of the memory buffer.