Added -c option to continue beyond suspicious code blocks
In most cases encountering 10+ NOPs or LD r,r instructions are signs
of a problem, but in a few cases these do exist in real titles.  The
new option allows a more complete trace, though in other cases it
would add too much non-code, so it's not enabled by default.

The Spanish title 3D Interceptor uses LD A,A instead of NOP as
part of a delay loop.  10th Frame (US Gold) includes a block of NOPs
that are programmatically poked with code.